KUDU-2191 (4/n): WIP: Hive Metastore catalog manager integration
This commit adds a basic integration with the Hive Metastore, which can
be enabled with a new flag: 'hive-metastore-addresses'. When the flag is
provided, the Master will push DDL changes in its catalog to the Hive
Metastore. In particular:

- Create table, delete table, and alter table operations in the Kudu
  master will synchronously update the corresponding entry in the HMS.

- New table and column names are required to be valid according to the
  Hive identifier rules, which are much stricter than Kudu's native
  identifier rules.

WIP: This patch doesn't yet have sufficient error handling logic, nor
test coverage of errors.  Here are the test cases I intend to add to
master_hms-itest before this is committed:

- Rename a Kudu table to a new name which already has an entry in the HMS
  (should fail).
- Rename a Kudu table when the original name doesn't have an entry in
  the HMS (should create an entry with the new name).
- Rename a Kudu table when the HMS is unreachable (should fail).
- Create a Kudu table with a name which already exists in the HMS
  (should fail).
- Create a Kudu table when the HMS is unreachable (should fail).
- Drop a Kudu table when the corresponding entry in the HMS does not
  exist (should succeed).
- Drop a Kudu table when the HMS is unreachable (should fail?).

Additionally, this patch as written overwrites the entire HMS table
entry during table rename operations. This is bad because it will wipe
out any stats that query engines like Impala have attached to the table.

Change-Id: I1b8f360408b05d6ac4f7359a8ac0dd889b196baf