rpc: add experimental rpc_reuseport flag
This option sets the SO_REUSEPORT socket option on a server's bound
socket. This socket option is supported on all platforms which Kudu
supports, except RHEL 6.4 and below.

The motivation is to use the option in the minicluster so that master
ports can be reserved by the mini cluster control processes and reused
by the master processes.

Change-Id: I5d8ce9faa646fa2be554f5cfdf8b6ed0c48b496e