mini-cluster: support parallel multi-master clusters
The external mini cluster previously required hard-coded ports for
multi-master clusters, which in practice meant that tests which require
multi-master clusters must be run serially.

This commit changes the external mini cluster to assign ports so that
many multi-master clusters can coexist in a single port-space at once.
To do this, the mini-cluster coordinator binds to N sockets with the
SO_REUSEPORT socket options, which effectively reserves the ports. Next,
it uses these reserved ports to start the masters. This method is also
advantageous in that in keeps the ports reserved while the masters may
be temporarily shutdown and restarted during the life of the
mini-cluster.

WIP: This commit removes the ability to set the desired master ports.
master_migration-itest relies on this functionality, so it's currently
broken. I'm not sure whether it would be better to restore this
functionality, or rewrite the test so that it doesn't need it.

Change-Id: I0b0ff7bfc179d8fdb1ed306d1bbd12acddeb060c