Extract connection retrying and HA support from HmsCatalog
Introduces a new abstraction, thrift::HaClient, for providing HA and
retrying support on top of Thrift clients like HmsClient and
SentryClient. The implementation is extracted from HmsCatalog, and is
not changed in any significant ways.

The implementation is inlined into a header because the class requires a
template parameter (to choose between HmsClient or SentryClient), and
forward declaring these instantiations is not possible because the
thrift module does not link to the hms or sentry modules.

No new tests are provided, since the class is tested by the existing
master_hms-itest test, and providing tests in the thrift module is not
possible because there are no concrete instantiations which can be done
without HmsClient or SentryClient.

This patch should have little or no functional changes.

Change-Id: Id8135f4c3995bba0ba28384c35696a1771ff5296